[TACHYON-1328] Reorganize commands in TfsShell
