+++ shell/src/main/java/tachyon/shell/command/LoadMetadataCommand.java
@@ -1,0 +1,58 @@
+/*
+ * Licensed to the University of California, Berkeley under one or more contributor license
+ * agreements. See the NOTICE file distributed with this work for additional information regarding
+ * copyright ownership. The ASF licenses this file to You under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the License. You may obtain a
+ * copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package tachyon.shell.command;
+
+import java.io.IOException;
+
+import tachyon.TachyonURI;
+import tachyon.client.file.TachyonFileSystem;
+import tachyon.client.file.options.LoadMetadataOptions;
+import tachyon.conf.TachyonConf;
+import tachyon.exception.TachyonException;
+
+/**
+ * Loads metadata for the given Tachyon path from UFS.
+ */
+public class LoadMetadataCommand extends AbstractTfsShellCommand {
+
+  public LoadMetadataCommand(TachyonConf conf, TachyonFileSystem tfs) {
+    super(conf, tfs);
+  }
+
+  @Override
+  public String getCommandName() {
+    return "loadMetadata";
+  }
+
+  @Override
+  protected int getNumOfArgs() {
+    return 1;
+  }
+
+  @Override
+  public void run(String... args) throws IOException {
+    TachyonURI inputPath = new TachyonURI(args[0]);
+
+    try {
+      LoadMetadataOptions recursive =
+          new LoadMetadataOptions.Builder().setRecursive(true).build();
+      mTfs.loadMetadata(inputPath, recursive);
+    } catch (TachyonException e) {
+      throw new IOException(e.getMessage());
+    }
+  }
+
+}
