+++ shell/src/main/java/tachyon/shell/command/ReportCommand.java
@@ -1,0 +1,51 @@
+/*
+ * Licensed to the University of California, Berkeley under one or more contributor license
+ * agreements. See the NOTICE file distributed with this work for additional information regarding
+ * copyright ownership. The ASF licenses this file to You under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the License. You may obtain a
+ * copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package tachyon.shell.command;
+
+import java.io.IOException;
+
+import tachyon.TachyonURI;
+import tachyon.client.file.TachyonFileSystem;
+import tachyon.client.lineage.TachyonLineageFileSystem;
+import tachyon.conf.TachyonConf;
+import tachyon.exception.TachyonException;
+
+/**
+ * Report to the master that a file is lost.
+ */
+public class ReportCommand extends WithWildCardPathCommand {
+
+  public ReportCommand(TachyonConf conf, TachyonFileSystem tfs) {
+    super(conf, tfs);
+  }
+
+  @Override
+  public String getCommandName() {
+    return "report";
+  }
+
+  @Override
+  void runCommand(TachyonURI path) throws IOException {
+    try {
+      TachyonLineageFileSystem.get().reportLostFile(path);
+      System.out.println(path + " has reported been reported as lost.");
+      CommandUtils.listLineages();
+    } catch (TachyonException e) {
+      throw new IOException(e.getMessage());
+    }
+  }
+
+}
