+++ shell/src/main/java/tachyon/shell/command/SetTTLCommand.java
@@ -1,0 +1,56 @@
+/*
+ * Licensed to the University of California, Berkeley under one or more contributor license
+ * agreements. See the NOTICE file distributed with this work for additional information regarding
+ * copyright ownership. The ASF licenses this file to You under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the License. You may obtain a
+ * copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package tachyon.shell.command;
+
+import java.io.IOException;
+
+import com.google.common.base.Preconditions;
+
+import tachyon.TachyonURI;
+import tachyon.client.file.TachyonFileSystem;
+import tachyon.conf.TachyonConf;
+
+/**
+ * Set a new TTL value for the file at path both of the TTL value and the path
+ * are specified by args.
+ */
+public class SetTTLCommand extends AbstractTfsShellCommand {
+
+  public SetTTLCommand(TachyonConf conf, TachyonFileSystem tfs) {
+    super(conf, tfs);
+  }
+
+  @Override
+  public String getCommandName() {
+    return "setTTL";
+  }
+
+  @Override
+  protected int getNumOfArgs() {
+    return 2;
+  }
+
+  @Override
+  public void run(String... args) throws IOException {
+    long ttlMs = Long.parseLong(args[1]);
+    Preconditions.checkArgument(ttlMs >= 0, "TTL value must be >= 0");
+    TachyonURI path = new TachyonURI(args[0]);
+    CommandUtils.setTTL(mTfs, path, ttlMs);
+    System.out.println("TTL of file '" + path + "' was successfully set to " + ttlMs
+        + " milliseconds.");
+  }
+
+}
