+++ shell/src/main/java/tachyon/shell/command/TfsShellCommand.java
@@ -1,0 +1,47 @@
+/*
+ * Licensed to the University of California, Berkeley under one or more contributor license
+ * agreements. See the NOTICE file distributed with this work for additional information regarding
+ * copyright ownership. The ASF licenses this file to You under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the License. You may obtain a
+ * copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package tachyon.shell.command;
+
+import java.io.IOException;
+
+/**
+ * A generalize interface for all the commands that can be run from TfsShell
+ */
+public interface TfsShellCommand {
+
+  /**
+   * Get the command name as input from the shell.
+   *
+   * @return the command name
+   */
+  String getCommandName();
+
+  /**
+   * Check if the arguments are valid.
+   *
+   * @param args the arguments for the command, excluding the command name
+   * @return whether the args are valid
+   */
+  boolean validateArgs(String... args);
+
+  /**
+   * Run the command.
+   *
+   * @param args the arguments for the command, excluding the command name
+   * @throws IOException
+   */
+  void run(String... args) throws IOException;
+}
