+++ shell/src/main/java/tachyon/shell/command/TouchCommand.java
@@ -1,0 +1,59 @@
+/*
+ * Licensed to the University of California, Berkeley under one or more contributor license
+ * agreements. See the NOTICE file distributed with this work for additional information regarding
+ * copyright ownership. The ASF licenses this file to You under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the License. You may obtain a
+ * copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package tachyon.shell.command;
+
+import java.io.IOException;
+
+import tachyon.TachyonURI;
+import tachyon.client.UnderStorageType;
+import tachyon.client.file.TachyonFileSystem;
+import tachyon.client.file.options.OutStreamOptions;
+import tachyon.conf.TachyonConf;
+import tachyon.exception.TachyonException;
+
+/**
+ * Creates a 0 byte file specified by argv. The file will be written to UnderFileSystem.
+ */
+public class TouchCommand extends AbstractTfsShellCommand {
+
+  public TouchCommand(TachyonConf conf, TachyonFileSystem tfs) {
+    super(conf, tfs);
+  }
+
+  @Override
+  public String getCommandName() {
+    return "touch";
+  }
+
+  @Override
+  protected int getNumOfArgs() {
+    return 1;
+  }
+
+  @Override
+  public void run(String... args) throws IOException {
+    TachyonURI inputPath = new TachyonURI(args[0]);
+
+    try {
+      mTfs.getOutStream(inputPath, new OutStreamOptions.Builder(mTachyonConf)
+          .setUnderStorageType(UnderStorageType.SYNC_PERSIST).build()).close();
+    } catch (TachyonException e) {
+      throw new IOException(e.getMessage());
+    }
+    System.out.println(inputPath + " has been created");
+  }
+
+}
